Block Library: Move Column resizing to Columns
